requirements.txt
# Core Bot Framework
python-telegram-bot[all]==20.7
python-dotenv==1.0.0

# HTTP & Async
aiohttp==3.9.1
requests==2.31.0

# Scheduling
apscheduler==3.10.4
pytz==2023.3

# Web Server (Health Checks)
flask==3.0.0

# Data & Feeds
feedparser==6.0.10

# AI/LLM (OpenRouter)
openrouter==0.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pylint==3.0.3

# Security
safety==2.3.5
cryptography==41.0.7

fastapi
sse-starlette
uvicorn
discord.py
